parent class for special cars
