+++ src/3d/body/eggplant-model.ts
@@ -4,10 +4,8 @@ import { Decal } from '../../model/decal';
 import { BodyTexture } from './body-texture';
 import { Body } from '../../model/body';
 import { PaintConfig } from '../../model/paint-config';
-import { getAssetUrl } from '../../utils/network';
 import { RocketConfig } from '../../model/rocket-config';
-import { bindColor, createTextureFromImage } from '../../utils/webgl';
-import { WebGLCanvasTexture } from '../../webgl/webgl-texture';
+import { PrimaryOnlyTextureWebGL } from '../../webgl/primary-only-texture-webgl';
 
 // language=GLSL
 const FRAGMENT_SHADER = `
@@ -44,87 +42,10 @@ const FRAGMENT_SHADER = `
 `;
 
 
-class EggplantBodySkin extends WebGLCanvasTexture implements BodyTexture {
-
-  protected fragmentShader: string = FRAGMENT_SHADER;
-
-  private readonly blankSkinUrl: string;
-
-  private bodyBlankSkin: HTMLImageElement;
-
-  private primary: Color;
-
-  private primaryLocation: WebGLUniformLocation;
-  private rgbaMapLocation: WebGLUniformLocation;
-
-  private rgbaMapTexture: WebGLTexture;
-
-  constructor(body: Body, paints: PaintConfig, rocketConfig: RocketConfig) {
-    super(getAssetUrl(body.base_skin, rocketConfig), rocketConfig);
-    this.blankSkinUrl = getAssetUrl(body.blank_skin, rocketConfig);
-    this.primary = paints.primary;
-  }
-
-  async load() {
-    const superTask = super.load();
-    const rgbaMapTask = this.loader.load(this.blankSkinUrl);
-
-    this.bodyBlankSkin = await rgbaMapTask;
-    await superTask;
-  }
-
-  protected initWebGL(width: number, height: number) {
-    this.rgbaMapLocation = this.gl.getUniformLocation(this.program, 'u_rgba_map');
-    this.primaryLocation = this.gl.getUniformLocation(this.program, 'u_primary');
-    super.initWebGL(width, height);
-  }
-
-  protected createTextures() {
-    super.createTextures();
-    this.rgbaMapTexture = createTextureFromImage(this.gl, this.bodyBlankSkin);
-  }
-
-  protected setTextureLocations() {
-    super.setTextureLocations();
-    this.gl.uniform1i(this.rgbaMapLocation, 1);
-  }
-
-  protected bindTextures() {
-    super.bindTextures();
-    this.gl.activeTexture(this.gl.TEXTURE1);
-    this.gl.bindTexture(this.gl.TEXTURE_2D, this.rgbaMapTexture);
-  }
-
-  protected update() {
-    bindColor(this.gl, this.primaryLocation, this.primary);
-    super.update();
-  }
-
-  dispose() {
-    super.dispose();
-    this.bodyBlankSkin = undefined;
-    this.gl.deleteTexture(this.rgbaMapTexture);
-  }
-
-  setAccent(color: Color) {
-  }
-
-  setBodyPaint(color: Color) {
-  }
-
-  setPaint(color: Color) {
-  }
-
-  setPrimary(color: Color) {
-    this.primary = color;
-    this.update();
-  }
-}
-
 export class EggplantModel extends BodyModel {
 
   initBodySkin(body: Body, decal: Decal, paints: PaintConfig, rocketConfig: RocketConfig): BodyTexture {
-    return new EggplantBodySkin(body, paints, rocketConfig);
+    return new PrimaryOnlyTextureWebGL(body, paints, rocketConfig, FRAGMENT_SHADER);
   }
 
   setPaintColor(color: Color) {
